app/requirements.txt
django==2.2.10
djangorestframework==3.12.4
psycopg2-binary==2.8.5
